+++ src/main/java/name/divinityunbound/item/ModItemGroups.java
@@ -54,6 +54,7 @@ public class ModItemGroups {
                         entries.add(ModItems.DIVINE_TORCH);
                         entries.add(ModItems.SAND_OF_TIME);
                         entries.add(ModItems.GOLD_BAND);
+                        entries.add(ModItems.BATTERY_CORE);
 
                         entries.add(ModFluids.SPACE_TIME_BUCKET);